+++ server/licenses/known/Apache-2.0.ts
@@ -1,2 +1,2 @@
-export const tags = ['notice', 'fixed-text'];
+export const tags = ['notice', 'fixed-text', 'popular'];
 export const text = true; // use SPDX version